Module Ir_gen.Ir_gen_env

val ir_gen_method_name : Ast.Ast_types.Method_name.t -> Ast.Ast_types.Class_name.t -> string

Name mangling of method names - takes as input the method name and the class of the object calling it and returns the mangled name

val ir_gen_field_index : Ast.Ast_types.Field_name.t -> Ast.Ast_types.Class_name.t -> Desugaring.Desugared_ast.class_defn list -> int

Given a field and the type of the object to which it belongs, and a list of class defns, get the field index within the list of field defns in the corresponding class defn

val get_class_fields : Ast.Ast_types.Class_name.t -> Desugaring.Desugared_ast.class_defn list -> Ast.Ast_types.field_defn list

val ir_gen_vtable : Ast.Ast_types.Class_name.t -> Desugaring.Desugared_ast.class_defn list -> string list

val ir_gen_vtable_method_index : Ast.Ast_types.Method_name.t -> Ast.Ast_types.Class_name.t -> Desugaring.Desugared_ast.class_defn list -> int