Module Ir_gen.Ir_gen_expr

val ir_gen_method_name : Ast.Ast_types.Method_name.t -> Ast.Ast_types.type_expr -> string Core.Or_error.t

Name mangling of method names - takes as input the method name and the type of the object calling it and returns the mangled name

val ir_gen_expr : Desugaring.Desugared_ast.expr -> Llvm_ast.expr Core.Or_error.t

Generates the serialisable IR from a desugared expr