Module Ir_gen__Ir_gen_class_and_function_defns

Generates serialisable IR representations of the desugared class and function definitions

val ir_gen_class_defns : Desugaring.Desugared_ast.class_defn list -> Ir_gen.Frontend_ir.class_defn list

val ir_gen_function_defns : Desugaring.Desugared_ast.class_defn list -> Desugaring.Desugared_ast.function_defn list -> Ir_gen.Frontend_ir.function_defn list

Generates IR function defns from the function defns and each class's method defns